+++ simulator/short-circuit/src/test/java/com/powsybl/incubator/simulator/shortcircuit/ShortCircuitNormTest.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) 2022, Jean-Baptiste Heyberger & Geoffroy Jamgotchian
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package com.powsybl.incubator.simulator.shortcircuit;
 
 import com.powsybl.iidm.network.*;
@@ -12,6 +18,9 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+/**
+ * @author Jean-Baptiste Heyberger <jbheyberger at gmail.com>
+ */
 public class ShortCircuitNormTest {
 
     @Test
